Delivery chute functional
The delivery chute field in the gui will not display whatever is in the
the delivery chute after any button has been pressed. This will also
remove all items from the delivery chute. I have also fixed the display
field so that it should only display what the display object has set as
a message